+++ services/api/src/offers/offers.module.ts
@@ -29,6 +29,7 @@ import { Offer } from './entities/offer.entity';
 import { OfferStateTransition } from './entities/offer-state-transition.entity';
 import { OfferDelivery } from './entities/offer-delivery.entity';
 import { User } from '../auth/entities/user.entity';
+import { AvailableOffersModule } from './available/available-offers.module';
 
 /**
  * Offers module.
@@ -56,6 +57,7 @@ import { User } from '../auth/entities/user.entity';
         defaultJobOptions: config.defaultJobOptions,
       }),
     ),
+    AvailableOffersModule,
   ],
   controllers: [OffersController],
   providers: [